Don't store system fields on the model
Switch strategy a bit.  Instead of storing system fields on the model,
just recognize them when accessing the entity fields and ignore.
