+++ corehq/apps/users/views/mobile/custom_data_fields.py
@@ -1,12 +1,12 @@
 from django.utils.decorators import method_decorator
 from django.utils.translation import ugettext as _
 
-from corehq.apps.custom_data_fields.views import CustomDataFieldsMixin
+from corehq.apps.custom_data_fields import CustomDataModelMixin
 from corehq.apps.users.decorators import require_can_edit_commcare_users
 from corehq.apps.users.views import BaseUserSettingsView
 
 
-class UserFieldsView(CustomDataFieldsMixin, BaseUserSettingsView):
+class UserFieldsView(CustomDataModelMixin, BaseUserSettingsView):
     urlname = 'user_fields_view'
     field_type = 'UserFields'
     entity_string = _("User")
